@@ -10,7 +10,7 @@
         }
     ],
     "require": {
-        "php": ">=5.5.9",
+        "php": ">=7.0",
         "illuminate/console": "~5.1",
         "illuminate/contracts": "~5.1",
         "illuminate/support": "~5.1",
@@ -18,7 +18,7 @@
         "symfony/var-dumper": "~3.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.0|~5.0"
+        "phpunit/phpunit": "~6.0"
     },
     "autoload": {
         "psr-4": {
